+++ src/blocks/blockAllContributors.test.ts
@@ -55,10 +55,16 @@ describe("blockAllContributors", () => {
 			  "files": {
 			    ".all-contributorsrc": "{
 			  "badgeTemplate": "\\t<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"👪 All Contributors: <%= contributors.length %>\\" src=\\"https://img.shields.io/badge/%F0%9F%91%AA_all_contributors-<%= contributors.length %>-21bb42.svg\\" /></a>",
+			  "commitType": "docs",
 			  "contributors": [],
+			  "contributorsPerLine": 7,
 			  "contributorsSortAlphabetically": true,
+			  "files": [
+			    "README.md"
+			  ],
 			  "projectName": "test-repository",
-			  "projectOwner": "test-owner"
+			  "projectOwner": "test-owner",
+			  "repoType": "github"
 			}",
 			    ".github": {
 			      "workflows": {
@@ -182,6 +188,7 @@ describe("blockAllContributors", () => {
 			  "files": {
 			    ".all-contributorsrc": "{
 			  "badgeTemplate": "\\t<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"👪 All Contributors: <%= contributors.length %>\\" src=\\"https://img.shields.io/badge/%F0%9F%91%AA_all_contributors-<%= contributors.length %>-21bb42.svg\\" /></a>",
+			  "commitType": "docs",
 			  "contributors": [
 			    {
 			      "avatar_url": "https://avatars.githubusercontent.com/u/3335181?v=4",
@@ -198,9 +205,14 @@ describe("blockAllContributors", () => {
 			      "profile": "http://www.joshuakgoldberg.com"
 			    }
 			  ],
+			  "contributorsPerLine": 7,
 			  "contributorsSortAlphabetically": true,
+			  "files": [
+			    "README.md"
+			  ],
 			  "projectName": "test-repository",
-			  "projectOwner": "JoshuaKGoldberg"
+			  "projectOwner": "JoshuaKGoldberg",
+			  "repoType": "github"
 			}",
 			    ".github": {
 			      "workflows": {
@@ -323,6 +335,7 @@ describe("blockAllContributors", () => {
 			  "files": {
 			    ".all-contributorsrc": "{
 			  "badgeTemplate": "\\t<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"👪 All Contributors: <%= contributors.length %>\\" src=\\"https://img.shields.io/badge/%F0%9F%91%AA_all_contributors-<%= contributors.length %>-21bb42.svg\\" /></a>",
+			  "commitType": "docs",
 			  "contributors": [
 			    {
 			      "avatar_url": "https://avatars.githubusercontent.com/u/3335181?v=4",
@@ -339,9 +352,14 @@ describe("blockAllContributors", () => {
 			      "profile": "http://www.example.com"
 			    }
 			  ],
+			  "contributorsPerLine": 7,
 			  "contributorsSortAlphabetically": true,
+			  "files": [
+			    "README.md"
+			  ],
 			  "projectName": "test-repository",
-			  "projectOwner": "test-owner"
+			  "projectOwner": "test-owner",
+			  "repoType": "github"
 			}",
 			    ".github": {
 			      "workflows": {