+++ src/blocks/blockMarkdownlint.test.ts
@@ -1,5 +1,5 @@
-import { testBlock } from "bingo-stratum-testers";
-import { describe, expect, test } from "vitest";
+import { testBlock, testIntake } from "bingo-stratum-testers";
+import { describe, expect, it, test } from "vitest";
 
 import { blockMarkdownlint } from "./blockMarkdownlint.js";
 import { optionsBase } from "./options.fakes.js";
@@ -68,9 +68,7 @@ describe("blockMarkdownlint", () => {
 			  ],
 			  "files": {
 			    ".markdownlint.json": "{"extends":"markdownlint/style/prettier","first-line-h1":false,"no-inline-html":false}",
-			    ".markdownlintignore": ".github/CODE_OF_CONDUCT.md
-			CHANGELOG.md
-			node_modules/
+			    ".markdownlintignore": "node_modules/
 			",
 			  },
 			  "scripts": [
@@ -151,9 +149,7 @@ describe("blockMarkdownlint", () => {
 			  ],
 			  "files": {
 			    ".markdownlint.json": "{"extends":"markdownlint/style/prettier","first-line-h1":false,"no-inline-html":false}",
-			    ".markdownlintignore": ".github/CODE_OF_CONDUCT.md
-			CHANGELOG.md
-			lib/
+			    ".markdownlintignore": "lib/
 			node_modules/
 			",
 			  },
@@ -244,9 +240,7 @@ describe("blockMarkdownlint", () => {
 			  ],
 			  "files": {
 			    ".markdownlint.json": "{"extends":"markdownlint/style/prettier","first-line-h1":false,"no-inline-html":false}",
-			    ".markdownlintignore": ".github/CODE_OF_CONDUCT.md
-			CHANGELOG.md
-			node_modules/
+			    ".markdownlintignore": "node_modules/
 			",
 			  },
 			  "scripts": [
@@ -259,5 +253,37 @@ describe("blockMarkdownlint", () => {
 			  ],
 			}
 		`);
+	});
+
+	describe("intake", () => {
+		it("returns undefined when .markdownlintignore does not exist", () => {
+			const actual = testIntake(blockMarkdownlint, {
+				files: {},
+			});
+
+			expect(actual).toBeUndefined();
+		});
+
+		it("returns no ignores when .markdownlintignore does not contain truthy lines", () => {
+			const actual = testIntake(blockMarkdownlint, {
+				files: {
+					".markdownlintignore": ["\n"],
+				},
+			});
+
+			expect(actual).toEqual({ ignores: [] });
+		});
+
+		it("returns ignores when .markdownlintignore contains lines", () => {
+			const ignores = ["abc", "def"];
+
+			const actual = testIntake(blockMarkdownlint, {
+				files: {
+					".markdownlintignore": [ignores.join("\n") + "\n"],
+				},
+			});
+
+			expect(actual).toEqual({ ignores });
+		});
 	});
 });