+++ src/blocks/blockReleaseIt.test.ts
@@ -13,6 +13,14 @@ describe("blockReleaseIt", () => {
 			  "addons": [
 			    {
 			      "addons": {
+			        "ignores": [
+			          "CHANGELOG.md",
+			        ],
+			      },
+			      "block": [Function],
+			    },
+			    {
+			      "addons": {
 			        "properties": {
 			          "devDependencies": {
 			            "@release-it/conventional-changelog": "10.0.0",
@@ -111,7 +119,7 @@ describe("blockReleaseIt", () => {
 			      - env:
 			          GITHUB_TOKEN: \${{ secrets.ACCESS_TOKEN }}
 			          NPM_TOKEN: \${{ secrets.NPM_TOKEN }}
-			        uses: JoshuaKGoldberg/release-it-action@v0.3.1
+			        uses: JoshuaKGoldberg/release-it-action@v0.3.2
 
 
 			name: Release
@@ -159,6 +167,14 @@ describe("blockReleaseIt", () => {
 		expect(creation).toMatchInlineSnapshot(`
 			{
 			  "addons": [
+			    {
+			      "addons": {
+			        "ignores": [
+			          "CHANGELOG.md",
+			        ],
+			      },
+			      "block": [Function],
+			    },
 			    {
 			      "addons": {
 			        "properties": {
@@ -262,7 +278,7 @@ describe("blockReleaseIt", () => {
 			      - env:
 			          GITHUB_TOKEN: \${{ secrets.ACCESS_TOKEN }}
 			          NPM_TOKEN: \${{ secrets.NPM_TOKEN }}
-			        uses: JoshuaKGoldberg/release-it-action@v0.3.1
+			        uses: JoshuaKGoldberg/release-it-action@v0.3.2
 
 
 			name: Release