+++ src/presets/common.ts
@@ -2,6 +2,7 @@ import { base } from "../base.js";
 import { blockAllContributors } from "../blocks/blockAllContributors.js";
 import { blockCodecov } from "../blocks/blockCodecov.js";
 import { blockFunding } from "../blocks/blockFunding.js";
+import { blockOctoGuide } from "../blocks/blockOctoGuide.js";
 import { blockReleaseIt } from "../blocks/blockReleaseIt.js";
 import { blockVitest } from "../blocks/blockVitest.js";
 import { presetMinimal } from "./minimal.js";
@@ -17,6 +18,7 @@ export const presetCommon = base.createPreset({
 		blockAllContributors,
 		blockCodecov,
 		blockFunding,
+		blockOctoGuide,
 		blockReleaseIt,
 		blockVitest,
 	],